LayoutTests/imported/w3c/web-platform-tests/webrtc-priority/RTCPeerConnection-ondatachannel-expected.txt


Harness Error (TIMEOUT), message = null

FAIL In-band negotiated channel created on remote peer should match the same configuration as local peer assert_equals: expected "high" but got "low"
TIMEOUT In-band negotiated channel created on remote peer should match the same (default) configuration as local peer Test timed out
